sqlbase: move time zone utils to package timeutil where they belong
